Stage 1/6: Making coffee
Description

Let's start with a program that makes you a coffee – virtual coffee, of course. In this project, you will implement functionality that simulates a real coffee machine. It can run out of ingredients, such as milk or coffee beans, it can offer you various types of coffee, and, finally, it will take money for the prepared drink.

Objective

The first version of the program just makes you a coffee. It should print to the standard output what it is doing as it makes the drink.